src/requirements.txt
# ######## MLFLOW / SQL data packages
# mlflow>=1.2.0
# catboost>=0.26
# boto3==1.17.90
python-dotenv==0.17.1

# ##### development
# jupyterlab
pytest 
sqlalchemy
httpx


# ##### Deep learning packages
# scikit-learn

######## Flask API packages
# flask
# flask-Cors
# flask-swagger-ui 
# pyyaml

######## Fast API packages
fastapi 
uvicorn
python-multipart
##### MySQL packages
mysql-connector-python


##### General packages
python-box==6.0.2
ensure==1.0.2
pandas
numpy
matplotlib
pyYAML
tqdm
joblib
types-PyYAML


## setup the project packages
-e .